requirements.txt
streamlit # for ui 
python-dotenv # for load secrets creditentials to main file
google-generativeai # for google gen ai